rename ProductPage to CartAppContainer, which now renders NavBar and ProductTable; App now renders CartAppContainer only; pass cart state down to NavBar and CartNavItem in order to display total product count on cart button in nav
