--- src/components/ProductPage.jsx
@@ -1,15 +1,0 @@
-import {useState} from 'react';
-import ProductTable from './ProductTable';
-
-const ProductPage = (props) => {
-  let [cart, setCart] = useState(0);
-
-  return (
-    <>
-      <p>{cart}</p>
-      <ProductTable setCart={setCart}/>
-    </>
-  )
-}
-
-export default ProductPage;
