+++ src/components/navBar.jsx
@@ -1,11 +1,11 @@
-import React from 'react';
 import CartNavItem from './CartNavItem';
 import '../styles/navBar.css';
 
-const NavBar = () => {
+const NavBar = (props) => {
+  console.log('navbar props: ', props)
   return (
     <div className="navBarWrapper">
-      <CartNavItem/>
+      <CartNavItem cart={props.cart}/>
     </div>
   )
 }
